ability to ignore fields on update (#11)
option to ignore field for updates

Often we will just use the update function with a case class. The goal is to NOT update fields that
are keys (like an id) because these shouldn't change and can cause deadlocks. I added a compile step
that will remove fields from update statements that are annotated to not update. Created wrapper profiles for mysql and h2 that include the compile step.